+++ usvg/tests/test.rs
@@ -18,7 +18,8 @@ macro_rules! test {
             let in_str = std::fs::read_to_string(format!("tests/files/{}-in.svg", name)).unwrap();
             let out_str = std::fs::read_to_string(format!("tests/files/{}-out.svg", name)).unwrap();
 
-            let tree = usvg::Tree::from_str(&in_str, &usvg::Options::default()).unwrap();
+            let opt = usvg::Options::default();
+            let tree = usvg::Tree::from_str(&in_str, &opt.to_ref()).unwrap();
 
             let xml_opt = usvg::XmlOptions {
                 id_prefix: None,
@@ -46,7 +47,7 @@ macro_rules! test_preserve {
                 keep_named_groups: true,
                 .. usvg::Options::default()
             };
-            let tree = usvg::Tree::from_str(&in_str, &re_opt).unwrap();
+            let tree = usvg::Tree::from_str(&in_str, &re_opt.to_ref()).unwrap();
 
             let xml_opt = usvg::XmlOptions {
                 id_prefix: None,
@@ -101,7 +102,8 @@ macro_rules! test_size {
         #[test]
         fn $name() {
             use usvg::FuzzyEq;
-            let tree = usvg::Tree::from_str($input, &usvg::Options::default()).unwrap();
+            let opt = usvg::Options::default();
+            let tree = usvg::Tree::from_str($input, &opt.to_ref()).unwrap();
             assert!(tree.svg_node().size.fuzzy_eq(&$expected));
         }
     };
@@ -126,7 +128,8 @@ macro_rules! test_size_err {
     ($name:ident, $input:expr) => {
         #[test]
         fn $name() {
-            assert!(usvg::Tree::from_str($input, &usvg::Options::default()).is_err());
+            let opt = usvg::Options::default();
+            assert!(usvg::Tree::from_str($input, &opt.to_ref()).is_err());
         }
     };
 }